requirements.txt
asgiref==3.11.1
charset-normalizer==3.4.6
dj-database-url==3.1.2
Django==6.0.3
djangorestframework==3.17.1
et_xmlfile==2.0.0
gunicorn==25.3.0
numpy==2.4.4
openpyxl==3.1.5
packaging==26.0
pandas==3.0.2
pillow==12.2.0
psycopg==3.3.3
psycopg-binary==3.3.3
python-dateutil==2.9.0.post0
reportlab==4.4.10
simplekml==1.3.6
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.12.0
